requirements.txt
# Core dependencies
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
PyYAML==6.0.1

# Computer vision
scikit-image==0.22.0
imutils==0.5.4

# Async support
aiofiles==23.2.1

# USB/HID support
pyusb==1.2.1
evdev==1.6.1

# Web API (optional)
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# CLI
click==8.1.7
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Raspberry Pi specific (install separately if on Pi)
# picamera2==0.3.12
# hailort  # Install via Hailo SDK